TYP: Concrete ``float64`` scalar type with ``builtins.float`` as a base class

+++ numpy/typing/tests/data/fail/arithmetic.pyi
@@ -10,7 +10,7 @@ td = np.timedelta64(0, "D")
 AR_b: npt.NDArray[np.bool]
 AR_u: npt.NDArray[np.uint32]
 AR_i: npt.NDArray[np.int64]
-AR_f: npt.NDArray[np.float64]
+AR_f: npt.NDArray[np.longdouble]
 AR_c: npt.NDArray[np.complex128]
 AR_m: npt.NDArray[np.timedelta64]
 AR_M: npt.NDArray[np.datetime64]
